+++ usvg/src/export.rs
@@ -863,8 +863,8 @@ fn write_path(
     xml.write_transform(AId::Transform, path.transform);
 
     xml.write_attribute_raw("d", |buf| {
-        for seg in path.data.iter() {
-            match *seg {
+        for seg in path.data.segments() {
+            match seg {
                 PathSegment::MoveTo { x, y } => {
                     buf.extend_from_slice(b"M ");
                     write_num(x, buf);